third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/xsl-img-blocked-expected.txt

CONSOLE ERROR: Refused to load the image 'http://127.0.0.1:8000/security/resources/abe.png' because it violates the following Content Security Policy directive: "img-src 'none'".

Here is an image: 
